fix: lgas table modified field

+++ database/migrations/2022_11_24_230100_create_lgas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('lgas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('state_id');
+            $table->unsignedBigInteger('state_id');
             $table->string('name');
             $table->timestamps();
         });
